+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Smooth the edges in `vis_drake_graph()` and `render_drake_graph()`.
 - Make hover text slightly more readable in in `vis_drake_graph()` and `render_drake_graph()`.
+- Improve `dependency_profile()` show major trigger hashes side-by-side
+to tell the user if the command, a dependency, an input file, or an ouptut file changed since the last `make()`.
 
 # Version 5.4.0